refactor: extract try_init_each helper from decode_many_from
